Implement SchafferN2 optimization problem, with code adjustments
A new file `SchafferN2.kt` is added for the Schaffer N2 optimization problem. This includes a main function for running genetic algorithms. Corresponding changes have been made across other files as required. Notably, adjustments in test files for the correct calculations, alteration of the 'Booth' example to use 'schaffer2', and changes in the 'Range' class to improve filter conditions for ranges. An enum 'ToStringMode' is also added to 'Domain.kt'.