Add new mutation tests and update keen-core version
New mutation scenarios are introduced in the mutation test. Furthermore, a comment for replacement with the latest version is added in the dependencies and the keen-core version is updated to 1.0.2. These additions help verify mutable conditions in genetic algorithms while the updates ensure the use of the recent library version.