Refactor MutatorTest and Alterers, adjust comment formatting in SchafferN2
This commit introduces several changes to the 'MutatorTest' and 'Alterers' classes related to the use of the 'anyMutator' function and the description of it. Besides, it corrects the formatting of the opening comment in 'SchafferN2' class. The primary change involves removing type parameters from the 'anyMutator' function and updating related tests.