Implement Genetic Algorithm examples for Traveling Salesman Problem
This commit includes the creation of classes 'RouteChromosome', 'RoutePointGene', 'FitnessFunction', 'Main', and 'TravelingSalesmanProblem' to showcase the application of the Genetic Algorithm on the Traveling Salesman Problem (TSP). Also, 'OrderedCrossover' and 'AbstractEvolutionListener' from the keen core have been updated for improved functionality. These examples contribute towards a deeper understanding and efficient application of the Genetic Algorithm on complex problems like TSP.