Update CHANGELOG, README, and Keen version
Added an explanation for the implementation of the "Travelling Salesman Problem Using Genetic Algorithm" in CHANGELOG.md. Revised and expanded the installation instructions in README.md and updated the Keen version in gradle.properties from 1.0.3 to 1.1.0. This update will help users understand the new feature and provides complete information on setting up the framework.