✨🔨✅ Add CrossoverInvocationException and update crossover functionality
This commit adds a new exception class, CrossoverInvocationException, to handle errors specifically related to the crossover operation in the genetic algorithm. The crossover functionality has been updated to throw this new exception when the number of parent genotypes doesn't match the number of parents. Additionally, minor tweaks to improve readability have been made, including a modification to the InversionMutator class. Tests for the new exception and updated crossover operation have also been added.