Refactor crossover function and add CrossoverResult interface
The crossover function in the Crossover.kt file has been refactored to provide a more robust error handling and validation system. The function now throws a CrossoverInvocationException for inconsistencies in the number of parent genotypes or variation in the number of chromosomes. The changes also include the addition of a new CrossoverResult interface. This interface serves to encapsulate the outcome of a crossover operation by providing information about the resultant genetic materials and the count of crossover events. Unit tests have also been updated to test the new functionalities.