✅🔧 Add equality constraint to crossover function
The crossover function has been updated to include a constraint that checks the equality of output size with the number of offspring. This constraint was added to ensure that the number of offspring produced matches the output size. Adjustments have also been made in the test file to accommodate this change.