✅🔧⚙️ Update AverageCrossoverTest; refactor CombineCrossover
This commit updates AverageCrossoverTest file to include new tests on "with gene rate 1.0" and "with arbitrary gene rate". It also reduces PropTestConfig iterations from 100 to 50. In the CombineCrossover.kt file, the gene combining logic has been refactored to handle cases of geneRate equals to 0.0 and 1.0 separately. The random geneRate case is also handled now.